posts/FullStack-整理前后端的交互方式/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
strawberry-graphql[fastapi]>=0.218.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
python-multipart>=0.0.6
